Rework constraints during type inference
Previously, constraints were accumulated and attached to each term in the tree recursively. In many cases this led to incorrect behaviour as terms had duplicate and unnecessary constraints.

Now, constraints are instead accumulated in a `WriterT`, and then discharged at the top-level once were done walking over a set of equations. This is pretty much the same approach taken by _Typing Haskell in Haskell_ except there they manually simulate a Writer Monad by passing a tuple of `(Constraint, Value)` as the return value.

A nice side effect of this change is that it actually makes the inference code a _little_ easier to read. There is still a lot more work to do in order for it to be presentable though.

I've added a regression for a bug caused by the previous behaviour.