Correctly instatiate the types of trait methods
The previous code would incorrectly instantiate trait methods as if
they were rank-1, which would bubble into more errors during desugaring.
For example the type of `fmap` is actually:

`forall f. forall a b. (a -> b) -> f a -> f b`

Previously the code would treat it as:

`forall f a b. (a -> b) -> f a -> f b`

This lays the groundwork to finally fix the desugaring of constrained methods.