Add a Renamer
For a long time, I had put off developing the renamer, but as I near
the 'finish line', it's time to start building out the module system
so I can build larger programs and libraries. The renamer processes
a raw module and resolves all the names inside of it.

Ironically, adding the renamer broke the interaction with builtin
functions and forced improve upon the previous setup for the `Prelude`
module where I simplify merged the contents of the prelude with
whatever_real_ module was being compiled.

This was also a good opportunity to flex the new error reporting
functionality of the compiler! I used it to report suggested fixes for
unknown names.